refactor(shared-utils): rearranging files prior to build/publish
